fix: cast severity constants to u32 to make it pass on windows

+++ crates/tracel-mlir-rs/src/diagnostic/severity.rs
@@ -13,19 +13,19 @@ pub enum DiagnosticSeverity {
     Warning,
 }
 
-// Cast bindgen constants to u32 so this works on targets where they're i32.
-const SEV_ERROR: u32 = MlirDiagnosticSeverity_MlirDiagnosticError;
-const SEV_NOTE: u32 = MlirDiagnosticSeverity_MlirDiagnosticNote;
-const SEV_REMARK: u32 = MlirDiagnosticSeverity_MlirDiagnosticRemark;
-const SEV_WARN: u32 = MlirDiagnosticSeverity_MlirDiagnosticWarning;
+// Normalize bindgen constants to u32 (bindgen may emit i32 on some targets).
+const SEV_ERROR:  u32 = MlirDiagnosticSeverity_MlirDiagnosticError  as u32;
+const SEV_NOTE:   u32 = MlirDiagnosticSeverity_MlirDiagnosticNote   as u32;
+const SEV_REMARK: u32 = MlirDiagnosticSeverity_MlirDiagnosticRemark as u32;
+const SEV_WARN:   u32 = MlirDiagnosticSeverity_MlirDiagnosticWarning as u32;
 
 #[inline]
 fn from_raw_u32(severity: u32) -> Result<DiagnosticSeverity, Error> {
     Ok(match severity {
-        SEV_ERROR => DiagnosticSeverity::Error,
-        SEV_NOTE => DiagnosticSeverity::Note,
+        SEV_ERROR  => DiagnosticSeverity::Error,
+        SEV_NOTE   => DiagnosticSeverity::Note,
         SEV_REMARK => DiagnosticSeverity::Remark,
-        SEV_WARN => DiagnosticSeverity::Warning,
+        SEV_WARN   => DiagnosticSeverity::Warning,
         _ => return Err(Error::UnknownDiagnosticSeverity(severity)),
     })
 }
